requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
azure-functions==1.6.0
certifi==2020.12.5
chardet==3.0.4
chromedriver-autoinstaller==0.2.2
grpcio==1.35.0
grpcio-tools==1.35.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.4
pandas==1.1.5
protobuf==3.14.0
pylint==2.6.0
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
tqdm==4.56.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
yarl==1.6.3
